chore: extract activity listener
